Interface and test implementation for policy.
The Xaya account registry contract will be non-upgradable, but it will
have a configurable link to a policy contract.  This can be used to
configure the policy for validating names and moves, as well as computing
and distributing the WCHI fee for those operations.

This commit adds the interface for the policy contract as well as an
implementation for use in unit tests.