Use StringBuilder for JSON/XML literals.
For constructing JSON or XML string literals in the NFT metadata
generation, we need to build up a string bit-by-bit.  Previously,
we were just using abi.encodePacked to concatenate the bits, which
was reallocating the string for each step (and leading to O(N^2)
complexity in the end).

This introduces a StringBuilder library, which can build up a string
for a pre-allocated maximum size in O(N) instead.