Return correct nonce from move function.
The "move" function from XayaAccounts is supposed to return the nonce
value that was used for the move, but it did actually return the next
nonce.  (But since the return value is used nowhere at the moment this
wasn't noticed anywhere.)

This fixes the function to properly return the actual nonce value used
(i.e. the same logged in the Move event).