Use data server instead of data: URIs.
It seems that OpenSea has some issues with generated data: URIs.  To work
around those, this introduces a "data server", which is just a centralised
URI.  NFT metadata and image URIs are pointing to this server instead of
using data: URIs directly.

The contract still contains all the logic for generating the metadata
and image as data: URIs, though, and this is what the data server will
actually serve.