Python script for the data server.
Implement a Python script and Docker container for the server that
can be queried for NFT metadata and images.  It will read it from
the on-chain generation functions and just forward the result.

SVG data is converted to PNG on the fly by the script as well, so that
we don't get issues with external image links.