Implement contractURI method.
This adds the contractURI method, which can return a link to contract
level metadata for OpenSea.  The main contract calls through to the
policy, and the initial policy calls the NftMetadata contract.

The metadata contract has just a variable with the link, which can be
configured by the owner at will.