Modernise NFT dataserver code.
Update the NFT dataserver code so it works better with a fresh build
of the Docker container.  We now need to set CairoSVG's "unsafe" flag to
get it to load external URLs, and we also changed the script so the contract
address is explicitly passed rather than loaded from the build artefact.  This
makes it easier to rebuild and run, without having to get the address
into the artefact file first.