Enable build by forge.
This restructures the repository to allow building the code with Foundry's
forge in addition to the legacy truffle.  Tests can still be run with
truffle and are not touched.

In order to do this, the OpenZeppelin dependency is converted from a node
module into a git submodule (at version 4.5.0 as used in the original code),
and the node dependency refers to the submodule locally.  The base64-sol
dependency is removed, and instead the Base64 util from OpenZeppelin is used.