Restrict namespace to exactly one character.
Restrict the namespace of newly registered names to be exactly one
character.  This is a work-around for a bug in XayaAccounts.tokenIdForName,
where the hashing is done incorrectly, so that namespace and name are just
put together, and thus e.g. "p/domob" and "pd/omob" map to the same token id.

With restrictions to exactly one character namespaces, this isn't an issue
anymore (for future names).