Fix hashing used to get token ID.
Fix the hashing structure used to convert (ns, name) pairs to a unique
token ID.  The previous structure is broken and can lead to (ns, name)
combinations that result in the same hash even though they are not the same
(but if we only use one-letter namespaces, such as in practice and now also
enforced on Polygon, then it works).

The new code properly handles arbitrary ns and name combinations.