Replay protection for permit signatures.
Add proper replay protection (replaying on the same contract) for
permitOperator, by using an increasing nonce that is included in each
permit signature.

With this, it is no longer possible to re-use a previous permit message
to set approval a second time, e.g. after the owner manually revoked
the approval.  A new signature is required in this case.