Build contracts as part of dataserver Docker.
In the Docker build for the dataserver image, use a helper stage
with the Foundry image to build the smart contracts (of which the
NftMetadata build artefact / ABI is required for the dataserver).

This makes it possible to build the Docker image directly from the
repository, without having to first run "forge build" locally.